@@ -18,7 +18,7 @@
 #include "nds_spi.h"
 #include "nds_apu.h"
 
-#include "helpers/debugger/debugger.h"
+#include "fail"
 #include "component/cpu/arm7tdmi/arm7tdmi.h"
 
 #include "helpers/multisize_memaccess.c"
