@@ -7,7 +7,7 @@
 #include <stdio.h>
 #include <string.h>
 
-#include "helpers/debugger/debugger.h"
+#include "fail"
 #include "helpers/physical_io.h"
 
 #include "tg16.h"
